rename processor example test files, update base64.decode doc
