+++ pkg/plugin/processor/builtin/impl/custom/z_examples_exporter_test.go
