+++ pkg/plugin/processor/builtin/impl/z_examples_exporter_test.go
